Removed possible collider bug
Right now we only changed the pointer on the array to point to nullptr. But the pointer to the collider on the module was still pointing to a position on the array. If we were to fill that position, the pointer in the module would be detected as filled and that could end up causing problems.